+++ lib/qc/modules/qc_tester/calculator/resources/calculator_service.dart
@@ -135,13 +135,15 @@ abstract class CalculatorService {
     return service.get("/manual-test/scan-device/$deviceBarcode", DeviceDetailResponse.fromJson);
   }
 
-  Stream<BaseActionResponse?> reportMismatch(List<String?> scannedImeiList, String deviceBarcode, String imeiImageUrl,
-      {String? timeoutReason}) {
+  Stream<BaseActionResponse?> reportMismatch(List<String?>? scannedImeiList, String deviceBarcode, String imeiImageUrl,
+      {String? timeoutReason, bool? isImei2Available, bool isAutoApproved = false}) {
     var req = {
       "imei": scannedImeiList,
       "qr": deviceBarcode,
       "image_url": imeiImageUrl,
       "rm": timeoutReason,
+      "iaa": isAutoApproved,
+      if (isImei2Available != null) "imna": isImei2Available,
     };
 
     return service.post("/device/mismatch/report/save", BaseActionResponse.fromJson, body: jsonEncode(req));