+++ lib/qc/modules/qc_tester/lob_devices/resources/device_detail_response.dart
@@ -39,8 +39,11 @@ class DeviceDetailResponseData {
   @JsonKey(name: "rm")
   Map<String, int>? reasons;
 
+  @JsonKey(name: "idia")
+  bool? isDeviceImeiApproved;
+
   DeviceDetailResponseData(
-      this.qrCode, this.imei1, this.serialNo, this.selectedCategoryId, this.categories, this.imei2, this.reasons);
+      this.qrCode, this.imei1, this.serialNo, this.selectedCategoryId, this.categories, this.imei2, this.reasons, this.isDeviceImeiApproved);
 
   static DeviceDetailResponseData fromJson(Map<String, dynamic> json) => _$DeviceDetailResponseDataFromJson(json);